Add config file
The following commit adds suport for config file. All specific details,
such as database name, or airline operator API URL will be added in this
file. The structure is the following: each group of data has a name and
multiple values with values. For example:
[DATABASE]
name = my_database

Values which do not belong to any group are automatically asigned to the
'DEFAULT' group.